+++ rotkehlchen/tests/unit/test_inquirer.py
@@ -719,12 +719,12 @@ def test_usd_price(inquirer: Inquirer, globaldb: GlobalDBHandler):
     """Check that price is queried for tokens in different chains using defillama"""
     inquirer.set_oracles_order(oracles=[CurrentPriceOracle.DEFILLAMA])
     globaldb.add_asset(EvmToken.initialize(
-        address=string_to_evm_address('0xDeadDeAddeAddEAddeadDEaDDEAdDeaDDeAD0000'),
+        address=string_to_evm_address('0x66a2A913e447d6b4BF33EFbec43aAeF87890FBbc'),
         chain_id=ChainID.BOBA,
         token_kind=EvmTokenKind.ERC20,
         decimals=18,
-        name='Wrapped Ether',
-        symbol='WETH',
+        name='USDC',
+        symbol='USDC',
     ))
     globaldb.add_asset(EvmToken.initialize(
         address=string_to_evm_address('0xB31f66AA3C1e785363F0875A1B74E27b85FD66c7'),
@@ -762,7 +762,7 @@ def test_usd_price(inquirer: Inquirer, globaldb: GlobalDBHandler):
         Asset('eip155:8453/erc20:0xd9aAEc86B65D86f6A7B5B1b0c42FFA531710b6CA'),  # base
         Asset('eip155:42161/erc20:0x5979D7b546E38E414F7E9822514be443A4800529'),  # arb
         Asset('eip155:43114/erc20:0xB31f66AA3C1e785363F0875A1B74E27b85FD66c7'),  # avax
-        Asset('eip155:288/erc20:0xDeadDeAddeAddEAddeadDEaDDEAdDeaDDeAD0000'),  # boba
+        Asset('eip155:288/erc20:0x66a2A913e447d6b4BF33EFbec43aAeF87890FBbc'),  # boba
         Asset('eip155:1101/erc20:0xA8CE8aee21bC2A48a5EF670afCc9274C7bbbC035'),  # zkevm
     ):
         price = inquirer.find_usd_price(token)